// src/api/api.ts
export interface LeaderBoardDto {
    /**
     * Место игрока в игре
     * @type {number}
     * @memberof LeaderBoardDto
     */
    place: number;
    /**
     * Баланс игрока
     * @type {number}
     * @memberof LeaderBoardDto
     */
    cash: number;
    /**
     * Имя игрока
     * @type {string}
     * @memberof LeaderBoardDto
     */
    name: string;
}
/**
 *
 * @export
 * @interface LeaderboardResponseDto
 */
export interface LeaderboardResponseDto {
    /**
     * Список игроков на текущей странице
     * @type {Array<LeaderBoardDto>}
     * @memberof LeaderboardResponseDto
     */
    players: Array<LeaderBoardDto>;
    /**
     * Общее количество завершивших игру игроков
     * @type {number}
     * @memberof LeaderboardResponseDto
     */
    totalPlayers: number;
    /**
     * Общее количество страниц
     * @type {number}
     * @memberof LeaderboardResponseDto
     */
    totalPages: number;
    /**
     * Текущая страница
     * @type {number}
     * @memberof LeaderboardResponseDto
     */
    currentPage: number;
    /**
     * Количество записей на странице
     * @type {number}
     * @memberof LeaderboardResponseDto
     */
    pageSize: number;
    /**
     * Есть ли следующая страница
     * @type {boolean}
     * @memberof LeaderboardResponseDto
     */
    hasNextPage: boolean;
    /**
     * Есть ли предыдущая страница
     * @type {boolean}
     * @memberof LeaderboardResponseDto
     */
    hasPreviousPage: boolean;
}
// ...
export interface GameApiGameLeaderboardRequest {
    /**
     * Номер страницы (начиная с 1)
     * @type {number}
     * @memberof GameApiGameLeaderboard
     */
    readonly page?: number;

    /**
     * Количество записей на странице
     * @type {number}
     * @memberof GameApiGameLeaderboard
     */
    readonly limit?: number;
}